+++ python/paddle/fluid/tests/unittests/test_pairwise_distance.py
@@ -110,6 +110,196 @@ class TestPairwiseDistance(unittest.TestCase):
         self.assertTrue(np.allclose(static_ret, excepted_value))
         self.assertTrue(np.allclose(dygraph_ret, excepted_value))
 
+#单测
+import unittest
+import paddle
+def call_pairwise_distance_layer(x, y, p=2., epsilon=1e-6, keepdim='False', name='name'):
+    pairwise_distance = paddle.nn.PairwiseDistance(
+        p=p, 
+        epsilon=epsilon, 
+        keepdim=keepdim, 
+        name=name)
+    distance = pairwise_distance(x=x, y=y)
+    return distance
+
+def call_pairwise_distance_functional(x, y, p=2., epsilon=1e-6, keepdim='False', name='name'):
+    distance = paddle.nn.functional.pairwise_distance(
+        x=x,
+        y=y,
+        p=p,
+        epsilon=epsilon,
+        keepdim=keepdim,
+        name=name)
+    return distance
+
+def test_static(place,
+                x_np,
+                y_np,
+                p=2,
+                epsilon=1e-6,
+                keepdim=False,
+                name='name',
+                functional=False):
+    prog = paddle.static.Program()
+    startup_prog = paddle.static.Program()
+    with paddle.static.program_guard(prog, startup_prog):
+        x = paddle.static.data(name='x',
+                               shape=x_np.shape,
+                               dtype='float64')
+        y = paddle.static.data(name='y',
+                               shape=y_np.shape,
+                               dtype='float64')   
+        feed_dict = {
+            "x": x,
+            "y": y,   
+        }   
+        if functional:
+            distance = call_pairwise_distance_functional(
+                x=x,
+                y=y,
+                p=p,
+                epsilon=epsilon,
+                keepdim=keepdim,
+                name=name)                     
+        else:
+            distance = call_pairwise_distance_layer(
+                x=x,
+                y=y,
+                p=p,
+                epsilon=epsilon,
+                keepdim=keepdim,
+                name=name)
+        exe = paddle.static.Executor(place)
+        static_result = exe.run(prog, feed=feed_dict, fetch_list=[distance])
+    return static_result
+
+def test_dygraph(place,
+                 x,
+                 y,
+                 p=2,
+                 epsilon=1e-6,
+                 keepdim=False,
+                 name='name',
+                 functional=False):
+    paddle.disable_static()
+    x = paddle.to_tensor(x)
+    y = paddle.to_tensor(y)
+
+    if functional:
+        dy_distance = call_pairwise_distance_functional(
+            x=x,
+            y=y,
+            p=p,
+            epsilon=epsilon,
+            keepdim=keepdim,
+            name=name)
+    else:
+        dy_distance = call_pairwise_distance_layer(
+            x=x,
+            y=y,
+            p=p,
+            epsilon=epsilon,
+            keepdim=keepdim,
+            name=name)
+    dy_distance = dy_distance.numpy()
+    paddle.enable_static()
+    return dy_distance
+
+def cala_pairwise_distance(x, y, p=2.0, epsilon=1e-6, keepdim=False, name='name'):
+    
+    distance = np.linalg.norm(x, y, p=p, epsilon=epsilon, keepdim=keepdim)
+
+    return distance
+
+class TestPairwiseDistance(unittest.TestCase):
+    def test_pairwisedistance(self):
+        shape = (3, 3)
+        x = np.random.uniform(0, 1, size=shape).astype(np.float64)
+        y = np.random.uniform(1, 2, size=shape).astype(np.float64)
+
+        places = [paddle.CPUPlace()]
+        keepdims = ['True', 'False']
+        p_list = [0, 1, 2, 'inf', '-inf']
+        if paddle.device.is_compiled_with_cuda():
+            places.append(paddle.CUDAPlace(0))
+        for place in places:
+            for p in p_list:
+                for keepdim in keepdims:
+                    expected = cala_pairwise_distance(
+                        x=x,
+                        y=y,
+                        p=p,
+                        keepdim=keepdim)
+
+                    dy_distance = test_dygraph(
+                        place=place,
+                        x=x,
+                        y=y,
+                        p=p,
+                        keepdim=keepdim
+                    )
+
+                    static_distance = test_static(
+                        place=place,
+                        x=x,
+                        y=y,
+                        p=p,
+                        keepdim=keepdim)
+                    self.assertTrue(np.allclose(static_distance, expected))
+                    self.assertTrue(np.allclose(static_distance, dy_distance))
+                    self.assertTrue(np.allclose(dy_distance, expected))
+
+                    static_functional = test_static(
+                        place=place,
+                        x=x,
+                        y=y,
+                        p=p,
+                        keepdim=keepdim,
+                        functional=True)
+                    dy_functional = test_dygraph(
+                        place=place,
+                        x=x,
+                        y=y,
+                        p=p,
+                        keepdim=keepdim,
+                        functional=True)
+                self.assertTrue(np.allclose(static_functional, expected))
+                self.assertTrue(np.allclose(static_functional, dy_functional))
+                self.assertTrue(np.allclose(dy_functional, expected))
+    def test_pairwise_distance_error(self):
+
+        paddle.disable_static()
+        self.assertRaises(ValueError,
+                        paddle.nn.PairwiseDistance,
+                        keepdim="unsupport keepdim")
+        x = paddle.to_tensor([[0.1, 0.3]], dtype='float32')
+        y = paddle.to_tensor([[0.0, 1.0]], dtype='float32')
+        self.assertRaises(
+            ValueError,
+            paddle.nn.functional.triplet_margin_with_distance_loss,
+            x=x,
+            y=y,
+            keepdim="unsupport keepdim")
+        
+        self.assertRaises(ValueError,
+                        paddle.nn.PairwiseDistance,
+                        p="unsupport keepdim")
+        x = paddle.to_tensor([[0.1, 0.3]], dtype='float32')
+        y = paddle.to_tensor([[0.0, 1.0]], dtype='float32')
+        self.assertRaises(
+            ValueError,
+            paddle.nn.functional.triplet_margin_with_distance_loss,
+            x=x,
+            y=y,
+            p="unsupport keepdim")
+        paddle.enable_static()
+
+    # def test_pairwise_distance_dimension(self):
+    #     paddle.disable_static()
+
+    #     x = paddle.to_tensor(np.random.randn(3, 4))
+    #     y = paddle.to_tensor(np.random.randn(4, ))
+    #     test_dygraph(x, y)
 
 if __name__ == "__main__":
     unittest.main()
